# src/main/java/com/ayurveda/ipr/chat/model/FivePillarsResponse.java
package com.ayurveda.ipr.chat.model;

import java.util.ArrayList;
import java.util.List;

/**
 * Structured 5-Pillars Response generated by IP-SHAKTI Orchestrator:
 * 1. 🛡️ IP Analysis (Patentability, Trade Marks, InPASS, USPTO)
 * 2. 📋 Regulatory Analysis (Rule 158-B, Form 25-D, Clinical Trials, AYUSH Evidence)
 * 3. 🌱 ABS / Biodiversity Check (NBA Form I/III, Section 3/6, Live UN CBD IRCCs)
 * 4. 📚 Prior-Art / TKDL Check (TKRC IPC classification, Sanskrit binomials, WIPO 1-click)
 * 5. 🔗 Official Statutory Sources (Clickable citations to the 27 PDFs & JSON datasets)
 */
public class FivePillarsResponse {

    private IpPillar ipAnalysis;
    private RegulatoryPillar regulatoryAnalysis;
    private AbsPillar absCheck;
    private TkdlPillar tkdlCheck;
    private List<StatutorySourceCitation> sources = new ArrayList<>();

    public FivePillarsResponse() {
    }

    public FivePillarsResponse(IpPillar ipAnalysis, RegulatoryPillar regulatoryAnalysis, 
                               AbsPillar absCheck, TkdlPillar tkdlCheck, 
                               List<StatutorySourceCitation> sources) {
        this.ipAnalysis = ipAnalysis;
        this.regulatoryAnalysis = regulatoryAnalysis;
        this.absCheck = absCheck;
        this.tkdlCheck = tkdlCheck;
        this.sources = sources;
    }

    // Getters and Setters
    public IpPillar getIpAnalysis() {
        return ipAnalysis;
    }

    public void setIpAnalysis(IpPillar ipAnalysis) {
        this.ipAnalysis = ipAnalysis;
    }

    public RegulatoryPillar getRegulatoryAnalysis() {
        return regulatoryAnalysis;
    }

    public void setRegulatoryAnalysis(RegulatoryPillar regulatoryAnalysis) {
        this.regulatoryAnalysis = regulatoryAnalysis;
    }

    public AbsPillar getAbsCheck() {
        return absCheck;
    }

    public void setAbsCheck(AbsPillar absCheck) {
        this.absCheck = absCheck;
    }

    public TkdlPillar getTkdlCheck() {
        return tkdlCheck;
    }

    public void setTkdlCheck(TkdlPillar tkdlCheck) {
        this.tkdlCheck = tkdlCheck;
    }

    public List<StatutorySourceCitation> getSources() {
        return sources;
    }

    public void setSources(List<StatutorySourceCitation> sources) {
        this.sources = sources;
    }

    // --- Sub-Pillar Classes ---

    public static class IpPillar {
        private boolean patentableInIndia;
        private String verdict;
        private List<String> relevantPatentSections;
        private String filingStrategy;
        private String inpassBooleanSyntax;
        private String usptoSection101Guidance;

        public IpPillar() {}

        public boolean isPatentableInIndia() { return patentableInIndia; }
        public void setPatentableInIndia(boolean patentableInIndia) { this.patentableInIndia = patentableInIndia; }
        public String getVerdict() { return verdict; }
        public void setVerdict(String verdict) { this.verdict = verdict; }
        public List<String> getRelevantPatentSections() { return relevantPatentSections; }
        public void setRelevantPatentSections(List<String> relevantPatentSections) { this.relevantPatentSections = relevantPatentSections; }
        public String getFilingStrategy() { return filingStrategy; }
        public void setFilingStrategy(String filingStrategy) { this.filingStrategy = filingStrategy; }
        public String getInpassBooleanSyntax() { return inpassBooleanSyntax; }
        public void setInpassBooleanSyntax(String inpassBooleanSyntax) { this.inpassBooleanSyntax = inpassBooleanSyntax; }
        public String getUsptoSection101Guidance() { return usptoSection101Guidance; }
        public void setUsptoSection101Guidance(String usptoSection101Guidance) { this.usptoSection101Guidance = usptoSection101Guidance; }
    }

    public static class RegulatoryPillar {
        private String productCategory;
        private String licensingAuthority;
        private String licensingForm;
        private String clinicalTrialObligation;
        private String governingActAndRules;
        private String ayushEvidenceChecklist;

        public RegulatoryPillar() {}

        public String getProductCategory() { return productCategory; }
        public void setProductCategory(String productCategory) { this.productCategory = productCategory; }
        public String getLicensingAuthority() { return licensingAuthority; }
        public void setLicensingAuthority(String licensingAuthority) { this.licensingAuthority = licensingAuthority; }
        public String getLicensingForm() { return licensingForm; }
        public void setLicensingForm(String licensingForm) { this.licensingForm = licensingForm; }
        public String getClinicalTrialObligation() { return clinicalTrialObligation; }
        public void setClinicalTrialObligation(String clinicalTrialObligation) { this.clinicalTrialObligation = clinicalTrialObligation; }
        public String getGoverningActAndRules() { return governingActAndRules; }
        public void setGoverningActAndRules(String governingActAndRules) { this.governingActAndRules = governingActAndRules; }
        public String getAyushEvidenceChecklist() { return ayushEvidenceChecklist; }
        public void setAyushEvidenceChecklist(String ayushEvidenceChecklist) { this.ayushEvidenceChecklist = ayushEvidenceChecklist; }
    }

    public static class AbsPillar {
        private String complianceStatus;
        private String requiredForm;
        private int activeIndianPermitsInCbd;
        private String liveCbdLink;
        private String governingSection;

        public AbsPillar() {}

        public String getComplianceStatus() { return complianceStatus; }
        public void setComplianceStatus(String complianceStatus) { this.complianceStatus = complianceStatus; }
        public String getRequiredForm() { return requiredForm; }
        public void setRequiredForm(String requiredForm) { this.requiredForm = requiredForm; }
        public int getActiveIndianPermitsInCbd() { return activeIndianPermitsInCbd; }
        public void setActiveIndianPermitsInCbd(int activeIndianPermitsInCbd) { this.activeIndianPermitsInCbd = activeIndianPermitsInCbd; }
        public String getLiveCbdLink() { return liveCbdLink; }
        public void setLiveCbdLink(String liveCbdLink) { this.liveCbdLink = liveCbdLink; }
        public String getGoverningSection() { return governingSection; }
        public void setGoverningSection(String governingSection) { this.governingSection = governingSection; }
    }

    public static class TkdlPillar {
        private String botanicalBinomial;
        private String sanskritName;
        private String tkrcIpcClass;
        private String priorArtRiskWarning;
        private String wipoPatentscopeDeepLink;

        public TkdlPillar() {}

        public String getBotanicalBinomial() { return botanicalBinomial; }
        public void setBotanicalBinomial(String botanicalBinomial) { this.botanicalBinomial = botanicalBinomial; }
        public String getSanskritName() { return sanskritName; }
        public void setSanskritName(String sanskritName) { this.sanskritName = sanskritName; }
        public String getTkrcIpcClass() { return tkrcIpcClass; }
        public void setTkrcIpcClass(String tkrcIpcClass) { this.tkrcIpcClass = tkrcIpcClass; }
        public String getPriorArtRiskWarning() { return priorArtRiskWarning; }
        public void setPriorArtRiskWarning(String priorArtRiskWarning) { this.priorArtRiskWarning = priorArtRiskWarning; }
        public String getWipoPatentscopeDeepLink() { return wipoPatentscopeDeepLink; }
        public void setWipoPatentscopeDeepLink(String wipoPatentscopeDeepLink) { this.wipoPatentscopeDeepLink = wipoPatentscopeDeepLink; }
    }
}